# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ExpenseItemPayload:
    properties:
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      description:
        type: string
      expense_date:
        type: string
      receipt_file_id:
        type: string
      source:
        type: string
      status:
        type: string
      tax_amount:
        type: number
      vendor_name:
        type: string
    type: object
  dto.ExpenseReportStatusPayload:
    properties:
      comment:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for the Business Trip and Expense Portal
  title: Travel Portal API
  version: "1.0"
paths:
...
  /expense-reports/{reportId}:
    get:
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get expense report by ID
      tags:
      - ExpenseReports
  /expense-reports/{reportId}/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Expense item payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseItemPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add expense item
      tags:
      - ExpenseReports
  /expense-reports/{reportId}/items/{itemId}:
    delete:
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete expense item
      tags:
      - ExpenseReports
    put:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: Expense item payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseItemPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update expense item
      tags:
      - ExpenseReports
  /expense-reports/{reportId}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseReportStatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change expense report status
      tags:
      - ExpenseReports
...
  /trip-requests/{tripId}/expense-report:
    get:
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get expense report for trip
      tags:
      - ExpenseReports
    post:
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create expense report
      tags:
      - ExpenseReports
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"